DEV: Simply removal of default loading spinner (#7)
Returning a falsey from the `loading()` action in the application route prevents the event from bubbling any further and triggering the 'loading substate' in the Ember router. If the loading substate is never hit, then the default loading template will never be displayed. This approach means we don't have to hack with the `Ember.TEMPLATES` object (which we are hoping to move away from in Discourse core).

In reality, we were (accidentally) already doing this. Calling `this._super()` in `loading()` on the Application route always returns an value of `undefined`. That passed our `!== null` check, and thereby returned a falsey value. The `return true` would never be hit.